% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_countries}
\alias{get_countries}
\title{Get Countries}
\usage{
get_countries(api_key = config::get("key"), parse_response = TRUE)
}
\arguments{
\item{api_key}{API key}

\item{parse_response}{Should we parse the response as a list?}
}
\value{
list of countries
}
\description{
List of the countries supported by the API.
}
\examples{
\dontrun{
get_countries()
}
}
