% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_currencies}
\alias{get_currencies}
\title{Get Currencies}
\usage{
get_currencies(api_key = config::get("key"), parse_response = TRUE)
}
\arguments{
\item{api_key}{API key}

\item{parse_response}{Should we parse the response as a list?}
}
\value{
list of currencies
}
\description{
List of the currencies supported by the API.
}
\examples{
\dontrun{
get_currencies()
}
}
