% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_keyword_data}
\alias{get_keyword_data}
\title{Get Keyword Data}
\usage{
get_keyword_data(
  keywords = c(),
  country = NULL,
  currency = NULL,
  data_source = NULL,
  api_key = config::get("key"),
  parse_response = TRUE
)
}
\arguments{
\item{keywords}{Vector of keywords, has to be fewer than 100.}

\item{country}{Country code also see \link[kwewr]{get_countries}}

\item{currency}{Currency code also see \link[kwewr]{get_currencies}}

\item{data_source}{"gkp" or "cli", NULL will default to cli}

\item{api_key}{API key}

\item{parse_response}{Should we parse the response as a list?}
}
\value{
list with volume, cpc, competion by keyword
}
\description{
Get Volume, CPC and competition for a set of keywords.
   For details consult [Keywords Everywhere API Docs](https://api.keywordseverywhere.com/docs/#/keywords/get_keywords_data)
}
\examples{
\dontrun{
get_keywords_data(c("keywords tool", "keyword planner"))
get_keywords_data(c("keywords tool", "keyword planner"), country = "us")
}
}
