% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwewr.R
\docType{package}
\name{kwewr-package}
\alias{kwewr}
\alias{kwewr-package}
\title{kwewr: R Keywords Everywhere API}
\description{
Simple interface for the Keywords Everywhere API.
}
\author{
\strong{Maintainer}: Reto Wyss \email{reto.wyss.ch@gmail.com}

}
\keyword{internal}
